fix: prevent initialization race condition between data module and tracker update
Itemchanger loads modules before it calls OnEnterGame. On save creation this is fine, as the tracker update module is added inside an OnEnterGame hook. After loading, however, the module is present to be loaded and tries to get data from the data module, which has not yet initialized. The bug is fixed by lazily fetching the data only when needed, which allows everything to finish initializing first.